Move colors file into utils
