+++ src/components/boards/board-column.tsx
@@ -2,7 +2,7 @@ import styled from "styled-components";
 import { Card } from "src/types/inner/card.g";
 import { ColumnCard } from "../columns/column-card";
 import { NewColumnCard } from "../columns/new-column-card";
-import { drelloColors } from "../../constants/colors";
+import { drelloColors } from "src/utils/colors";
 
 // Styled Components
 const ColumnMain = styled.div`
